+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "drawer-count-app",
-  "version": "0.2.3",
+  "version": "0.2.4",
   "description": "A minimal, installable PWA to help myself and coworkers count drawers at the end of the day.",
   "main": "index.html",
   "scripts": {
